resume_parser_langchain/requirements.txt
langchain>=0.3.0
langchain-openai>=0.3.0
pydantic>=2.8.0
python-dotenv>=1.0.1
